MCUIO: shield drivers, set TIAN gpios for i2c devices
This patch is temporary and is needed to start using Lucky shield connected
to Tian board. Mapping of gpios has been replaced from those for Yun to those
for Tian. So, with this patch, shield drivers will not work for Yun anymore.

The definitive solution is to use the new gpiolib interface which permits
to define a gpio mapping based on the functionalities and abstracting it
from gpio number.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>